+++ .env.example
@@ -34,6 +34,7 @@ JWT_SECRET=change-this-secret-key
 
 # 개인 API Key 입력
 GEMINI_API_KEY=your-api-key-here
+GEMINI_API_URL=gemini.api.url:https://generativelanguage.googleapis.com/v1beta/models/gemini-3.0-flash:generateContent
 AI_API_KEY=your-api-key-here
